Tests/LibWeb/Text/expected/css/computed-values-svg-geometry.txt

rx: 12px
ry: 34px
stroke-dasharray: 1px, 2
stroke-width: 25%
